Feat: Infra 레이어 작성 및 domain 과 infra 분리
