+++ src/member/repository/member.repository.interface.ts
@@ -1,1 +1,8 @@
+import { IRepository } from "@app/database/repository.interface";
 import { Member } from "@prisma/client";
+
+export interface IMemberRepository extends IRepository<Member> {
+  updateBalance(id: number, balance: number): Promise<Member>;
+}
+
+export const IMEMBER_REPOSITORY = Symbol("IMEMBER_REPOSITORY");
