+++ src/payment/repository/payment.repository.interface.ts
@@ -1,1 +1,16 @@
+import { IRepository } from "@app/database/repository.interface";
 import { Payment } from "@prisma/client";
+import { PaymentStatus } from "../domain/dto/payment-status.enum";
+
+export interface IPaymentRepository extends IRepository<Payment> {
+  createPayment(
+    orderId: number,
+    couponId: number,
+    memberId: number,
+    paid_amount: number,
+    approved_at: Date,
+    status: PaymentStatus,
+  ): Promise<Payment>;
+}
+
+export const IPAYMENT_REPOSITORY = Symbol("IPAYMENT_REPOSITORY");
