+++ src/product/domain/service/product.service.spec.ts
@@ -1,12 +1,12 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { ProductRepository } from "../infrastructure/product.repository";
+import { ProductRepository } from "../../infrastructure/product.repository";
 import { ProductService } from "./product.service";
 import { Product } from "../entity/product.entity";
 import { ProductResult } from "../dto/product.result";
 import { GetProductCommand } from "../dto/get-product.command";
 import { AddStockCommand } from "../dto/add-stock.command";
 import { DeductStockCommand } from "../dto/deduct-stock.command";
-import { IPRODUCT_REPOSITORY } from "../product.repository.interface";
+import { IPRODUCT_REPOSITORY } from "../../repository/product.repository.interface";
 
 describe("ProductService", () => {
   let productService: ProductService;
