+++ src/product/domain/service/product.service.ts
@@ -3,9 +3,9 @@ import { ProductResult } from "../dto/product.result";
 import { GetProductCommand } from "../dto/get-product.command";
 import { AddStockCommand } from "../dto/add-stock.command";
 import { DeductStockCommand } from "../dto/deduct-stock.command";
-import { ProductRepository } from "../infrastructure/product.repository";
+import { ProductRepository } from "../../infrastructure/product.repository";
 import { Product } from "@prisma/client";
-import { IPRODUCT_REPOSITORY } from "../product.repository.interface";
+import { IPRODUCT_REPOSITORY } from "../../repository/product.repository.interface";
 
 @Injectable()
 export class ProductService {
